REACT+FAST_API_Authentication/backend/requirements.txt
bcrypt
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib